scaffoldy/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-crispy-forms
django-query
django-debug-toolbar
gunicorn==20.1.0
requests
Jinja2
gitpython
django-bootstrap-modal-forms
django-widget-tweaks
django-ipware
django-extensions
ipython
#django-star-ratings